improve tape listing and tape trap handling

+++ emulation/interface.h
@@ -245,8 +245,8 @@ struct Interface {
 	// general purpose emulator output listing
 	struct Listing {
 		unsigned id;
-		std::vector<uint8_t> line; // host is responsible for conversion, e.g. c64 use petscii charset
-		std::vector<uint8_t> loadCommand;
+		std::vector<uint16_t> line; // host is responsible for conversion, e.g. c64 use petscii charset
+		std::vector<uint16_t> loadCommand;
 	};	
     
     enum Region : uint8_t { Pal = 0, Ntsc = 1 };
